back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "django>=4.2.4,<5",
    "django-cors-headers>=4.2.0",
    "django-environ>=0.10.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
    "djoser==2.3.3",
    "drf-spectacular==0.28.0",
    "psycopg2-binary>=2.9.10",
]
